refactor: Rename vorabcheck folder to shared
