+++ app/routes/beratungshilfe.vorabcheck.ergebnis.$.tsx
@@ -1,1 +1,1 @@
-export { action, loader, Step as default } from "~/routes/vorabcheck/result";
+export { action, loader, Step as default } from "~/routes/shared/result";
