+++ app/routes/geld-einklagen.vorabcheck.$.tsx
@@ -1,1 +1,1 @@
-export { action, loader, Step as default } from "~/routes/vorabcheck/step";
+export { action, loader, Step as default } from "~/routes/shared/step";
